Refaktoreringer
